osint-harvester/requirements.txt
requests
pandas
praw
rich
python-dotenv
GitPython
scikit-learn
numpy



google-generativeai
flask
